requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
filelock==3.14.0
grpcio==1.64.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
MouseInfo==0.1.3
openai==1.30.5
protobuf==4.25.3
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.7.2
pydantic_core==2.18.3
PyGetWindow==0.0.9
pyht==0.0.28
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.1
pytweening==1.2.0
requests==2.32.3
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
